requirements.txt
Django==5.0.0
django-debug-toolbar==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.23.0
Pillow==11.0.0
flake8==6.0.0
flake8-quotes==3.3.2
flake8-variables-names==0.0.6
pep8-naming==0.13.3
psycopg2-binary==2.9.10
psycopg2==2.9.10
drf_yasg
django-rest-swagger
python-dotenv